+++ doc/source/doxygen-docs/changelog.md
@@ -1,5 +1,9 @@
 \page changelog Change Log
 
+# Version 2.1.6: (Under development)
+- (No changes yet).
+
+------
 # Version 2.1.5: Released Dec 6th, 2020
 - Changes in libraries:
   - \ref mrpt_containers_grp